authorized/requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
Django==4.2.7
django-cors-headers==4.3.1
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
google-auth==2.25.2
google-oauth==1.0.1
gunicorn==21.2.0
idna==3.4
mysqlclient==2.2.0
packaging==23.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
waitress==2.1.2
